requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

numpy==1.26.4
opencv-python==4.10.0.84
scikit-image==0.21.0
pillow==9.5.0
onnx==1.19.1
onnxruntime-gpu==1.23.2
protobuf==4.25.1
psutil==6.0.0
packaging==24.1
PySide6==6.8.2.1 
kornia
torch==2.10.0+cu128
torchvision==0.25.0+cu128
torchaudio==2.10.0+cu128
tensorrt==10.8.0.43 --extra-index-url https://pypi.nvidia.com
tensorrt-cu12_libs==10.8.0.43
tensorrt-cu12_bindings==10.8.0.43
tqdm
ftfy
regex
pyvirtualcam==0.11.1
numexpr
onnxsim==0.4.36
requests
pyqt-toast-notification==1.3.2
qdarkstyle
pyqtdarktheme
